libvnme: Remove	__nvme_msg from export map
nvme-cli was using the library infrastructure for logging. In the
meantime this has been removed, hence we can remove this function from
the export map.

Furthermore, move nvme_msg() and __nvme_msg() to private header.
